fix: map shows all points + COALESCE type error in dispatch signatures
- Map: geocodeAddress returns {coords, exact} — non-geocoded points spread
  in Fibonacci spiral so all are visible (not stacked at same coordinate)
- Map: approximate markers shown in amber with ~ badge and warning popup
- COALESCE types text/integer: cast id::integer in dispatch_signatures_pending INSERT
- Route search filter: remove route.name TS error (already in prev commit)

+++ backend/controllers/dispatch.controller.ts
@@ -57,7 +57,7 @@ export const initDispatch = async (req: Request, res: Response) => {
 
             await pool.query(`
                 INSERT INTO dispatch_signatures_pending (id, dispatch_id, user_id, role_type, signed, signed_at)
-                SELECT COALESCE(MAX(id),0)+1, $1, $2, $3, $4, $5
+                SELECT COALESCE(MAX(id::integer),0)+1, $1, $2, $3, $4, $5
                 FROM dispatch_signatures_pending
             `, [dispatchId, sig.userId, sig.role, isSigned, signedAt]);
         }
